Parse any input as a document. Fix #8
Context: https://github.com/KirillOsenkov/XmlParser/issues/8

This commit fix the issue by adhering more closely to the XML grammar i.e. have any input be parsed as if it was a document at the top-level and fixing that process by making the presence of a prologue optional.

It also updates some unit-tests to be valid XML (aka have a root node).