Support retrieving SkippedTokensTriviaSyntax
"Flatten" the structured trivia such as SkippedTokensTriviaSyntax into a new SyntaxTrivia that just has the full string of the trivia.
Not quite sure this is what is supposed to happen but it seems to somewhat be working. The downside is that we allocate the SyntaxTrivia on every access or enumeration, perhaps worth caching it somehow later. But for now this is better than crashing.

Adds a test that ensures that if we enumerate all tokens and concatenate their trivia and text we get the exact original text back.

Fixes #59